Clean up package
