+++ package.json
@@ -24,13 +24,11 @@
   },
   "homepage": "https://github.com/RyanCavanaugh/idembot#readme",
   "dependencies": {
-    "@types/async": "^2.0.40",
     "@types/commander": "^2.9.1",
-    "@types/mkdirp": "^0.3.29",
-    "@types/moment": "^2.13.0",
-    "@types/mz": "0.0.30",
-    "@types/node": "^7.0.38",
-    "@types/request": "0.0.41",
+    "@types/mkdirp": "^0.5.0",
+    "@types/mz": "^0.0.31",
+    "@types/node": "^8.0.22",
+    "@types/request": "^2.0.1",
     "@types/winston": "^2.3.3",
     "bragi": "^0.1.3",
     "commander": "^2.11.0",
