+++ src/logging.ts
@@ -1,12 +1,12 @@
-import * as bragi from 'bragi';
+// import * as bragi from 'bragi';
 
 export interface Log {
     log(message: string, data?: any): void;
 }
 
-export function get(groupName: string): Log {
+export function get(_groupName: string): Log {
     return {
-        log: (message: string, data?: any) =>  console.log(message)// bragi.log(message, data)
+        log: (message: string, data?: any) =>  console.log(message, data)// bragi.log(message, data)
     };
 }
