--- src/modules.d.ts
@@ -1,4 +1,0 @@
-declare module "sleep-promise" {
-    function sleep(milliseconds: number): Promise<void>;
-    export = sleep;
-}
