+++ src/types/sleep-promise.d.ts
@@ -1,0 +1,2 @@
+declare function sleep(milliseconds: number): Promise<void>;
+export = sleep;
