+++ tsconfig.json
@@ -1,20 +1,21 @@
 {
-    "files": [
-        "src/index.ts",
-        "src/run.ts",
-        "src/bragi.d.ts",        
-        "src/modules.d.ts"
+    "include": [
+        "src"
     ],
     "compilerOptions": {
         "outDir": "bin",
         "module": "commonjs",
         "target": "es5",
-        "lib": ["es2015", "es5"],
-        "noImplicitAny": true,
-        "noImplicitThis": true,
-        "noImplicitReturns": true,
-        "strictNullChecks": true,
+        "lib": ["es2015"],
+        "strict": true,
+        "noUnusedLocals": true,
+        "noUnusedParameters": true,
         "sourceMap": false,
-        "declaration": true
+        "declaration": true,
+
+        "baseUrl": ".",
+        "paths": {
+            "*": ["src/types/*"]
+        }
     }
 }
